% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{a_m_no}
\alias{a_m_no}
\title{DAISIE datalist of spiders on Hawaii (Big Island) - a_m_no}
\format{
A list with 53 elements. The first one contains the \verb{$island_age},
a numeric with the island age in Ma, and \verb{$not_present}, a numeric with the
number of known species on the archipelago not present on Big Island.

Each subsequent element of the list is comprised of 5 elements:
\describe{
\item{$colonist_name}{character vector with name of each lineage.}
\item{$branching_times}{A decreasingly sorted numeric vector. The first
element is always the island age. Subsequent elements are either the
speciation times or information regarding the minimum age at which the
lineage could have colonised the archipelago.}
\item{$stac}{A numeric integer from 1 to 9. Defines the status of the
colonist, which has implications for the maximum likelihood inference.
See \link[DAISIE:DAISIE_dataprep]{DAISIE::DAISIE_dataprep} for more details.}
\item{$mising_species}{A numeric with the known number of missing species
that are not present in the phylogeny, or cannot otherwise be assigned
a branching time.}
\item{$type1or2}{A numeric defining if the clade is of type 1 or 2, for 2
type analyses. Always set to 1 in this dataset.}

}
}
\source{
Jairo Patiño and Rosemary Gilliespie. Full reference list to be
filled in later.
}
\usage{
a_m_no
}
\description{
A DAISIE datalist as required by \code{\link[DAISIE:DAISIE_ML]{DAISIE::DAISIE_ML()}}. Generated by running
\code{\link[DAISIE:DAISIE_dataprep]{DAISIE::DAISIE_dataprep()}} on \link{a_m_no_datatable}. This
dataset favours cladogenesis and does not assume MaxAge for most clades.
MinAge is never used when precise colonisation time is known. Island age is
set to 3.6My.
}
\references{
GARB, J.E. and GILLESPIE, R.G. (2009), Diversity despite dispersal:
colonization history and phylogeography of Hawaiian crab spiders inferred
from multilocus genetic data. Molecular Ecology, 18: 1746-1764.
\doi{10.1111/j.1365-294X.2009.04125.x}
}
\keyword{datasets}
