% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_output_folder.R
\name{create_output_folder}
\alias{create_output_folder}
\title{Create model output folder}
\usage{
create_output_folder(output_name, model)
}
\arguments{
\item{output_name}{A string. Will be the name of the created output folder.}

\item{model}{Numeric integer with model that should run}
}
\value{
Creates appropriate directory. Returns string with path for
output object.
}
\description{
Create model output folder
}
\examples{
\dontrun{
create_output_folder(output_name = "results_folder", model = 1)
}
}
\author{
Pedro Neves
}
