% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_params_doc.R
\name{default_params_doc}
\alias{default_params_doc}
\title{Default parameter documentation}
\usage{
default_params_doc(island_replicates)
}
\arguments{
\item{island_replicates}{List with DAISIE simulation output as generated by
functions such as \code{DAISIE::DAISIE_sim_constant_rate()}.}
}
\value{
Nothing
}
\description{
Default parameter documentation
}
\author{
Pedro Neves
}
\keyword{internal}
