% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_loglik_plot.R
\name{generate_loglik_plot}
\alias{generate_loglik_plot}
\title{Plot loglikelihood through time}
\usage{
generate_loglik_plot(res)
}
\arguments{
\item{res}{A data frame with each row corresponding to the best of 10 model
fits with a different set of randomly sampled starting parameters.}
}
\value{
A plot of loglik through time.
}
\description{
Plot loglikelihood through time
}
\examples{
data(res_lsodes)
plot <- generate_loglik_plot(res_lsodes)
}
