% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hawaiispiders-package.R
\docType{package}
\name{hawaiispiders-package}
\alias{hawaiispiders}
\alias{hawaiispiders-package}
\title{hawaiispiders: Analyse Spider Evolution on Hawaii Island}
\description{
Using the DAISIE framework, runs macroevolutionary analyses on the macroevolution of spider lineages on the Hawaii island, also known as Big Island. Explicitly accounts for the geodynamics of Hawaii.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Neves-P/hawaiispiders}
  \item Report bugs at \url{https://github.com/Neves-P/hawaiispiders/issues}
}

}
\author{
\strong{Maintainer}: Pedro Santos Neves \email{p.m.santos.neves@rug.nl} (\href{https://orcid.org/0000-0003-2561-4677}{ORCID})

}
\keyword{internal}
