% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{m_master}
\alias{m_master}
\title{Master DAISIE datatable of spiders on Hawaii (Big Island) favouring
cladogenesis}
\format{
A tibble with 52 rows and 5 variables:
\describe{
\item{Clade_name}{character vector with name of each lineage}
\item{Status}{character vector with endemicity status. Can be
\code{"Endemic"}, or \code{"NonEndemic"}. the \verb{_MaxAge} and \verb{_MaxAgeMinAge} suffixes
are not present in the master files (such as this one) and are added by
\code{make_data_set_file.R} for each of the scenarios \code{r_m_no}... etc.}
\item{Missing_species}{numeric with the number of known species that are
missing in the phylogeny}
\item{Branching_times}{ordered numeric vector with colonisation and
branching times of each lineage. Can also contain \code{"MinAges"} when these
are available}
}
}
\source{
Jairo Patiño and Rosemary Gilliespie. Full reference list to be
filled in later.
}
\usage{
m_master
}
\description{
A master dataset containing the colonization and branching times of spiders
on Big Island. This is then read and parsed by the script in
\code{data-raw/make_data_set_file.R} which appends the \verb{_MaxAge} and
\verb{_MaxAgeMinAge} suffixes given each scenario's uncertainty treatment and
island age. This parsing generates the actual scenarios used in the study,
such as \code{"r_m_no_datatable"}, \code{"r_m_no"}, etc. Endemicity status refers
strictly to Big Island endemics, and not endemics to the Hawaiian archipelago
in general. This master dataset favours migration.
}
\references{
GARB, J.E. and GILLESPIE, R.G. (2009), Diversity despite dispersal:
colonization history and phylogeography of Hawaiian crab spiders inferred
from multilocus genetic data. Molecular Ecology, 18: 1746-1764.
\doi{10.1111/j.1365-294X.2009.04125.x}
}
\keyword{datasets}
