% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_faceted_plot.R
\name{make_faceted_plot}
\alias{make_faceted_plot}
\title{Title}
\usage{
make_faceted_plot(
  plotting_table,
  parameter_names = get_parameter_names(),
  x_partition_by,
  colour_by,
  shape_by
)
}
\value{

}
\description{
Title
}
