% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_res_df.R
\name{make_res_df}
\alias{make_res_df}
\title{Make results data frame}
\usage{
make_res_df(sens)
}
\value{
A data frame with the results and the corresponding scenario
}
\description{
Make results data frame
}
\author{
Pedro Santos Neves
}
