% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_results_to_plot.R
\name{prepare_results_to_plot}
\alias{prepare_results_to_plot}
\title{Prepare raw results table into plottable data}
\usage{
prepare_results_to_plot(raw_results)
}
\value{
A data frame consiting of:
}
\description{
Prepare raw results table into plottable data
}
