% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_main_header.R
\name{print_main_header}
\alias{print_main_header}
\title{Print session and run info to console/log file}
\usage{
print_main_header(datalist_name, M, model)
}
\arguments{
\item{datalist_name}{Character with name of datalist to be used. Should be
an exported dataset found in \verb{/data} accessible by \code{\link[=data]{data()}}}

\item{M}{Numeric integer with mainland species number. In this case it refers
to the total number of species on the Hawaiian archipelago. Defaults to 165}

\item{model}{Numeric integer with model that should run}
}
\value{
Nothing. Prints session and run info to console.
}
\description{
Print session and run info to console/log file
}
\examples{
\dontrun{
print_main_header(
  datalist_name = "spider_scenario_1_1_1_datalist",
  M = 165,
  model = 1
)
}
}
\author{
Pedro Neves
}
