% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{res_lsodes}
\alias{res_lsodes}
\title{Results of MLE procedure on spider data yielding DAISIE parameter estimates}
\format{
A data frame with 36 rows and 10 variables:
\describe{
\item{scenario}{character vector with name of each scenario}
\item{lambda_c}{numeric atomic vector with estimated value of cladogenesis
rate}
\item{mu}{numeric atomic vector with estimated value of extinction
rate}
\item{K}{numeric atomic vector with estimated value of carrying capacity
}
\item{gamma}{numeric atomic vector with estimated value of colonisation
rate}
\item{lambda_a}{numeric atomic vector with estimated value of anagenesis
rate}
\item{loglik}{numeric atomic vector with calculated log-likelihood value of
the model fit.}
\item{df}{numeric atomic vector with the degrees of freedom in the model}
\item{conv}{numeric atomic vector indicating if the optimisation procedure
converged. A value of \code{0} denotes convergence.}
\item{bic}{numeric atomic vector with the calculated value of the Bayesian
Information Criterion}
}
}
\source{
See accompanying publication.
}
\usage{
res_lsodes
}
\description{
A data frame obtained by running the file \code{data-raw/res_lsodes} on all sets
of MLE procedures that yielded likelihood values and DAISIE parameter
estimates. Only the best fit as chosen by BIC for each scenario is kept in
the resulting data frame. Further details can be found in the document ion
for \code{DAISIEutils::sensitivity()}. All MLE procedures used the \code{lsodes}
numerical integrator from  \code{R} package \code{deSolve} from within the \code{DAISIE}
package.
}
\keyword{datasets}
