% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_main.R
\name{run_main}
\alias{run_main}
\title{Run analysis from a set of initial parameters}
\usage{
run_main(datalist_name, M = 165, model, seed_range)
}
\arguments{
\item{datalist_name}{Character with name of datalist to be used. Should be
an exported dataset found in \verb{/data} accessible by \code{\link[=data]{data()}}}

\item{M}{Numeric integer with mainland species number. In this case it refers
to the total number of species on the Hawaiian archipelago. Defaults to 165}

\item{model}{Numeric integer with model that should run}

\item{seed_range}{Numeric vector of 1 or 2. If only one element is supplied
then the function will sample initial parameters with \code{seed = seed_range}.
If the vector contains two elements, then \code{seed_range[2] - seed_range[1]}
analysis will run in succession, with seeds ranging from
\code{seed_range[1]:seed_range[2]}}
}
\value{
Saves results to .txt file.
}
\description{
Run analysis from a set of initial parameters
}
\examples{
# Run 1 set of randomly sampled initial parameters for scenario 1 and
# model 1
\dontrun{
run_main(
  datalist_name = "spider_scenario_1_1_1_datalist",
  M = 165,
  model = 1,
  seed_range = 1
)
}
}
\author{
Pedro Neves
}
