% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{run_spider_analysis}
\alias{run_spider_analysis}
\title{Run spider analysis}
\usage{
run_spider_analysis(datalist_name, M = 165, model, seed)
}
\arguments{
\item{datalist_name}{Character with name of datalist to be used. Should be
an exported dataset found in \verb{/data} accessible by \code{\link[=data]{data()}}}

\item{M}{Numeric integer with mainland species number. In this case it refers
to the total number of species on the Hawaiian archipelago. Defaults to 165}

\item{model}{Numeric integer with model that should run}

\item{seed}{Integer to be used as seed}
}
\value{
Nothing. Writes to file.
}
\description{
Run spider analysis
}
\examples{
\dontrun{
run_spider_analysis(
  datalist_name = "spider_scenario_1_datalist",
  M = 165,
  model = 1,
  seed = 1
)
}
}
