% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_model_parameters.R
\name{setup_model_arguments}
\alias{setup_model_arguments}
\title{Title}
\usage{
setup_model_arguments(model, r_lamc, r_mu, r_K, r_gam, r_ana)
}
\arguments{
\item{variables}{}
}
\value{

}
\description{
Title
}
