% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spider_scenario_1_pruned}
\alias{spider_scenario_1_pruned}
\title{DAISIE datatable of spiders on Hawaii (Big Island)}
\format{
A tibble with 31 rows and 5 variables:
\describe{
\item{Clade_name}{character vector with name of each lineage}
\item{Status}{character vector with endemicity status. Can be \code{"Endemic"},
\code{"NonEndemic"} or \code{"Endemic_MaxAge"}}
\item{Missing_species}{numeric with the number of known species that are
missing in the phylogeny}
\item{Branching_times}{ordered numeric vector with colonisation and
branching times of each lineage}
}
}
\source{
Jairo Patiño and Rosemary Gilliespie. Full reference list to be
filled in later.
}
\usage{
spider_scenario_1_pruned
}
\description{
A dataset containing the colonization and branching times of spiders on Big
Island. This format can be passed to \code{\link[DAISIE:DAISIE_dataprep]{DAISIE::DAISIE_dataprep()}}. Endemicity
status refers strictly to Big Island endemics, and not endemics to the
Hawaiian archipelago in general. Temporary file to be removed. Does not
include branching times older than island age on cladogenetic lineages.
}
\references{
GARB, J.E. and GILLESPIE, R.G. (2009), Diversity despite dispersal:
colonization history and phylogeography of Hawaiian crab spiders inferred
from multilocus genetic data. Molecular Ecology, 18: 1746-1764.
\url{https://doi.org/10.1111/j.1365-294X.2009.04125.x}
}
\keyword{datasets}
