% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spider_scenario_2_1_1_datatable}
\alias{spider_scenario_2_1_1_datatable}
\title{DAISIE datalist of spiders on Hawaii (Big Island) - 1.2.3}
\format{
A list with 34 elements. The first one contains the \verb{$island_age},
a numeric with the island age in Ma, and \verb{$not_present}, a numeric with the
number of known species on the archipelago not present on Big Island.

Each subsequent element of the list is comprised of 5 elements:
\describe{
\item{$colonist_name}{character vector with name of each lineage.}
\item{$branching_times}{A decreasingly sorted numeric vector. The first
element is always the island age. Subsequent elements are either the
speciation times or information regarding the minimum age at which the
lineage could have colonised the archipelago.}
\item{$stac}{A numeric integer from 1 to 9. Defines the status of the
colonist, which has implications for the maximum likelihood inference.
See \link[DAISIE:DAISIE_dataprep]{DAISIE::DAISIE_dataprep} for more details.}
\item{$mising_species}{A numeric with the known number of missing species
that are not present in the phylogeny, or cannot otherwise be assigned
a branching time.}
\item{$type1or2}{A numeric defining if the clade is of type 1 or 2, for 2
type analyses. Always set to 1 in this dataset.}

}

A tibble with 35 rows and 5 variables:
\describe{
\item{Clade_name}{character vector with name of each lineage}
\item{Status}{character vector with endemicity status. Can be
\code{"Endemic_MaxAge"}, \code{"NonEndemic_MaxAge"}, \code{"Endemic_MaxAgeMinAge"} or
\code{"NonEndemic_MaxAgeMinAge"}}
\item{Missing_species}{numeric with the number of known species that are
missing in the phylogeny}
\item{Branching_times}{ordered numeric vector with colonisation and
branching times of each lineage. Can also contain \code{"MinAges"} when these
are available}
}
}
\source{
Jairo Patiño and Rosemary Gilliespie. Full reference list to be
filled in later.

Jairo Patiño and Rosemary Gilliespie. Full reference list to be
filled in later.
}
\usage{
spider_scenario_2_1_1_datatable
}
\description{
A DAISIE datalist as required by \code{\link[DAISIE:DAISIE_ML]{DAISIE::DAISIE_ML()}}. Generated by running
\code{\link[DAISIE:DAISIE_dataprep]{DAISIE::DAISIE_dataprep()}} on \link{spider_scenario_1_2_3_datatable}. This
dataset favours cladogenesis and assumes MaxAge for most clades. MinAge is
never used when precise colonisation time is known. Island age is set to
2.4My.
}
\references{
GARB, J.E. and GILLESPIE, R.G. (2009), Diversity despite dispersal:
colonization history and phylogeography of Hawaiian crab spiders inferred
from multilocus genetic data. Molecular Ecology, 18: 1746-1764.
\doi{10.1111/j.1365-294X.2009.04125.x}
DAISIE datatable of spiders on Hawaii (Big Island) - 2.1.1

A dataset containing the colonization and branching times of spiders on Big
Island. This format can be passed to \code{\link[DAISIE:DAISIE_dataprep]{DAISIE::DAISIE_dataprep()}}. Endemicity
status refers strictly to Big Island endemics, and not endemics to the
Hawaiian archipelago in general. This dataset favours immigration and
assumes MaxAge for most clades. Island age is set to 1.2My.

GARB, J.E. and GILLESPIE, R.G. (2009), Diversity despite dispersal:
colonization history and phylogeography of Hawaiian crab spiders inferred
from multilocus genetic data. Molecular Ecology, 18: 1746-1764.
\doi{10.1111/j.1365-294X.2009.04125.x}
}
\keyword{datasets}
