% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spider_scenarios.R
\name{spider_scenarios}
\alias{spider_scenarios}
\title{List all spider scenarios}
\usage{
spider_scenarios()
}
\value{
Character vector, with all spider scenarios in the study, such as
\code{"y_c_max"}.
}
\description{
List all spider scenarios
}
\examples{
scenarios <- spider_scenarios()
}
\author{
Pedro Santos Neves
}
