% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_parameter_name.R
\name{translate_parameter_name}
\alias{translate_parameter_name}
\title{Title}
\usage{
translate_parameter_name(parameter_name)
}
\arguments{
\item{parameter_name}{A string with the code name of 1 parameter. Possible
names are returned by \code{get_parameter_names()} (although only 1 is accepted).}
}
\value{

}
\description{
Title
}
