% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_partition_name.R
\name{translate_partition_name}
\alias{translate_partition_name}
\title{Translate partitioning codes into plottable English}
\usage{
translate_partition_name(partition_name)
}
\value{
A character string with formatted name of partition
}
\description{
Translate partitioning codes into plottable English
}
\examples{
plottable_name <- translate_partition_name(partition_name = "island_age")
}
