% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stt.R
\name{convert_to_classic_plot}
\alias{convert_to_classic_plot}
\title{Prepare input for stt}
\usage{
convert_to_classic_plot(simulation_outputs)
}
\value{
a list with wrangled data to be used for plotting STT plots with
plot_stt
}
\description{
Prepare input for stt
}
\note{
This function is taken from \code{\link[DAISIE:DAISIE-package]{DAISIE::DAISIE-package}} and simplified, as it
is not exported there.
}
\examples{
utils::data("islands_1type_1000reps", package = "DAISIE")
simulation_outuputs <- hawaiispiders:::convert_to_classic_plot(
  islands_1type_1000reps
)

}
\keyword{internal}
