% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_output_folder.R
\name{create_output_folder}
\alias{create_output_folder}
\title{Create model output folder}
\usage{
create_output_folder(output_name, model, verbose = FALSE)
}
\arguments{
\item{output_name}{A string. Will be the name of the created output folder.}

\item{model}{Numeric integer with model that should run}

\item{verbose}{Boolean. Defaults to FALSE. If TRUE, session info and run
parameters are printed to the console/log file.}
}
\value{
Creates appropriate directory. Returns string with path for
output object. If on Peregrine, folder will be $HOME/results/$output_name.
If called from another environment, folder will be
getwd()/results/$output_name.
}
\description{
Create model output folder
}
\examples{
\dontrun{
create_output_folder(output_name = "results_folder", model = 1)
}
}
\author{
Pedro Neves
}
