% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_params_doc.R
\name{default_params_doc}
\alias{default_params_doc}
\title{Default parameters documentation}
\usage{
default_params_doc(model, r_lamc, r_mu, r_K, r_gam, r_ana)
}
\arguments{
\item{model}{Integer with model that should run}

\item{r_lamc}{Uniformly sampled value of cladogenesis rate}

\item{r_mu}{Uniformly sampled value of extinction rate}

\item{r_K}{Uniformly sampled value of carrying capacity}

\item{r_gam}{Uniformly sampled value of immigration rate}

\item{r_ana}{Uniformly sampled value of anagenesis rate}
}
\value{
Nothing
}
\description{
Default parameters documentation
}
