% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_params_doc.R
\name{default_params_doc}
\alias{default_params_doc}
\title{Default parameter documentation}
\usage{
default_params_doc(
  island_replicates,
  plotting_table,
  parameter_names,
  partition_by,
  colour_by,
  shape_by,
  sens,
  parameter_name,
  raw_results,
  plot_to_save,
  name,
  type_size,
  res
)
}
\arguments{
\item{island_replicates}{List with DAISIE simulation output as generated by
functions such as \code{DAISIE::DAISIE_sim_constant_rate()}.}

\item{plotting_table}{A \code{data.frame} as returned by
\code{prepare_results_to_plot()}. It contains the full results of the package
estimate runs together with necessary columns for data partitioning. Has 14
columns, the first 10 of which are as returned by
\code{DAISIEutils::sensitivity()} and the remaining four contain the
integration, island age, cladogenesis/migration grouping and treatment of
uncertainty grouping. All these are made up as factors to allow easy
plotting.}

\item{parameter_names}{Vector with names of parameters using the codes
usually found on \code{DAISIE} and as returned by \code{get_parameter_names()} by
default. Can be of length 1 to 5, allows for:
\itemize{
\item \code{"lac"}
\item \code{"mu"}
\item \code{"K"}
\item \code{"gam"}
\item \code{"laa"}
}}

\item{partition_by}{A string indicating what the x-axis should be
partitioned by. Options are \code{island_age} for partitioning by assumed island
age, \code{stac} for partitioning by the treatment of uncertainty chosen, and
\code{c_m} for partitioning by cladogenesis/migration favouring strategy.}

\item{colour_by}{A string indicating what data partitioning should be used
for colours. Options are \code{island_age} for partitioning by assumed island
age, \code{stac} for partitioning by the treatment of uncertainty chosen, and
\code{c_m} for partitioning by cladogenesis/migration favouring strategy.}

\item{shape_by}{A string indicating what data partitioning should be used
for shapes. Options are \code{island_age} for partitioning by assumed island
age, \code{stac} for partitioning by the treatment of uncertainty chosen, and
\code{c_m} for partitioning by cladogenesis/migration favouring strategy.}

\item{sens}{A data frame consisting of the output of \code{DAISIE::sensitivity()},
which can include several scenarios.}

\item{parameter_name}{A string with the code name of 1 parameter. Possible
names are returned by \code{get_parameter_names()} (although only 1 is
accepted).}

\item{raw_results}{A data frame with the model results as returned by
\code{make_res_df()}.}

\item{plot_to_save}{A ggplot which will be saved to disk.}

\item{type_size}{A string determining whether the figure should take full
size in a page: \code{"full_size"} or only half: \code{"half_size"}}

\item{res}{A data frame with each row corresponding to the best of 10 model
fits with a different set of randomly sampled starting parameters.}
}
\value{
Nothing
}
\description{
Default parameter documentation
}
\author{
Pedro Santos Neves
}
\keyword{internal}
