% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_params_doc.R
\name{default_params_doc}
\alias{default_params_doc}
\title{Default parameters documentation}
\usage{
default_params_doc(
  model,
  r_lamc,
  r_mu,
  r_K,
  r_gam,
  r_ana,
  datalist_name,
  M,
  seed,
  seed_range,
  file_path,
  output_name,
  verbose
)
}
\arguments{
\item{model}{Numeric integer with model that should run}

\item{r_lamc}{Numeric. Uniformly sampled value of cladogenesis rate}

\item{r_mu}{Numeric. Uniformly sampled value of extinction rate}

\item{r_K}{Numeric. Uniformly sampled value of carrying capacity}

\item{r_gam}{Numeric. Uniformly sampled value of immigration rate}

\item{r_ana}{Numeric. Uniformly sampled value of anagenesis rate}

\item{datalist_name}{Character with name of datalist to be used. Should be
an exported dataset found in \verb{/data} accessible by \code{\link[=data]{data()}}}

\item{M}{Numeric integer with mainland species number. In this case it refers
to the total number of species on the Hawaiian archipelago. Defaults to 165}

\item{seed}{Integer to be used as seed}

\item{seed_range}{Numeric vector of 1 or 2. If only one element is supplied
then the function will sample initial parameters with \code{seed = seed_range}.
If the vector contains two elements, then \code{seed_range[2] - seed_range[1]}
analysis will run in succession, with seeds ranging from
\code{seed_range[1]:seed_range[2]}}

\item{file_path}{The system directory where the output files will be stored.}

\item{output_name}{A string. Will be the name of the created output folder.}

\item{verbose}{Boolean. Defaults to FALSE. If TRUE, session info and run
parameters are printed to the console/log file.}
}
\value{
Nothing
}
\description{
Default parameters documentation
}
