% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_loglik_plot.R
\name{generate_loglik_plot}
\alias{generate_loglik_plot}
\title{Plot loglikelihood through time}
\usage{
generate_loglik_plot(res)
}
\arguments{
\item{res}{}
}
\value{
A plot of loglik through time.
}
\description{
Plot loglikelihood through time
}
\examples{
data(res_odeint)
plot <- generate_loglik_plot(res_odeint)
}
