% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_estimate_results.R
\name{plot_estimate_results}
\alias{plot_estimate_results}
\title{Title}
\usage{
plot_estimate_results(
  plotting_table,
  parameter_name,
  x_partition_by,
  colour_by,
  shape_by
)
}
\value{
A plot of one estimate partitioned by a given factor
}
\description{
Title
}
