% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sims.R
\name{plot_sims}
\alias{plot_sims}
\title{Plot island species-through-time (STT) plots}
\usage{
plot_sims(island_replicates)
}
\arguments{
\item{island_replicates}{List with DAISIE simulation output as generated by
functions such as \code{DAISIE::DAISIE_sim_constant_rate()}.}
}
\value{
R plot.
}
\description{
Produces STT plots.

R plots with number of total, endemic and non-endemic STTs for different
types of species for the entire time span the islands were simulated.
2.5-97.5th percentiles are plotted in light green, 25-75th percentiles
plotted in dark green.
}
\note{
This is an alternative colour scheme with minor modifications from
\code{\link[DAISIE:DAISIE_plot_sims]{DAISIE::DAISIE_plot_sims()}}. This version is simplified and always assumes
simulations were ran with \code{sample_freq} other than \code{Inf} and
}
\examples{
### Plot islands with single process (only one type of species)
utils::data(islands_1type_1000reps, package = "DAISIE")
plot_sims(
  island_replicates = islands_1type_1000reps
)

}
\references{
Valente, L.M., A.B. Phillimore and R.S. Etienne (2015).
Equilibrium and non-equilibrium dynamics simultaneously operate in the
Galapagos islands. Ecology Letters 18: 844-852.
}
\author{
Luis Valente, Pedro Neves
}
\keyword{models}
