% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stt.R
\name{plot_stt}
\alias{plot_stt}
\title{Create the Species-Through-Time plot. This is used to visualize
the output of sim functions}
\usage{
plot_stt(time, plot_lists = plot_lists, type = type)
}
\description{
Create the Species-Through-Time plot. This is used to visualize
the output of sim functions
}
\keyword{internal}
