% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_results_to_plot.R
\name{prepare_results_to_plot}
\alias{prepare_results_to_plot}
\title{Prepare raw results table into plottable data}
\usage{
prepare_results_to_plot(raw_results)
}
\arguments{
\item{raw_results}{A data frame with the model results as returned by
\code{make_res_df()}.}
}
\value{
A data frame containing the full results of the package
estimate runs together with necessary columns for data partitioning. Has 14
columns, the first 10 of which are as returned by
\code{DAISIEutils::sensitivity()} and the remaining four contain the
integration, island age, cladogenesis/migration grouping and treatment of
uncertainty grouping. All these are made up as factors to allow easy
plotting.
}
\description{
Prepare raw results table into plottable data
}
\examples{
\dontrun{
full_res <- readRDS("full_res.rds")
results_to_plot <- prepare_results_to_plot(full_res)
}
}
