% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_paper_plot.R
\name{save_paper_plot}
\alias{save_paper_plot}
\title{Save a plot to current directory}
\usage{
save_paper_plot(plot_to_save, type_size = "full_size", name)
}
\value{
NULL. A pdf and png files with the plot to save are saved to the
working directory
}
\description{
Save a plot to current directory
}
