% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_partition_name.R
\name{translate_partition_name}
\alias{translate_partition_name}
\title{Translate partitioning codes into plottable English}
\usage{
translate_partition_name(partition_name)
}
\value{

}
\description{
Translate partitioning codes into plottable English
}
