% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_partition_name.R
\name{translate_partition_name}
\alias{translate_partition_name}
\title{Translate partitioning codes into plottable English}
\usage{
translate_partition_name(partition_name)
}
\arguments{
\item{partition_name}{A string indicating with data partitioning codes.
Options are \code{island_age} for partitioning by assumed island
age, \code{stac} for partitioning by the treatment of uncertainty chosen, and
\code{c_m} for partitioning by cladogenesis/migration favouring strategy.}
}
\value{
A character string with formatted name of partition
}
\description{
Translate partitioning codes into plottable English
}
\examples{
plottable_name <- translate_partition_name(partition_name = "island_age")
}
