% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_partition_name.R
\name{translate_partition_name}
\alias{translate_partition_name}
\title{Translate partitioning codes into plottable English}
\usage{
translate_partition_name(partition_name)
}
\arguments{
\item{partition_name}{A string corresponding to the internal code for
partition names. Can be \code{"island_age"} for "Assumued island age (Ma)",
\code{"c_m"} for "Uncertainty in number of colonisations", or \code{"stac"} for
"Uncertainty in colonisation time"}
}
\value{
A character string with formatted name of partition
}
\description{
Translate partitioning codes into plottable English
}
\examples{
plottable_name <- translate_partition_name(partition_name = "island_age")
}
